requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
colorthief==0.2.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pylint==2.6.0
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
urllib3==1.26.2
webcolors==1.11.1
wrapt==1.12.1
